Moved plotting convenience functions to separate file.
